+++ web-ui/src/components/calendar-components/MobileCalendar.tsx
@@ -1,4 +1,4 @@
-import { Paper, Table, TableBody, TableCell, TableContainer, TableHead, TableRow } from '@mui/material';
+import { Paper, Table, TableBody, TableCell, TableContainer, TableHead, TableRow, styled } from '@mui/material';
 import { formatToShort } from '../../utils/sim-time-utils';
 import { useTranslation } from 'react-i18next';
 import { CalendarDto } from '../../../generated';
@@ -9,6 +9,13 @@ interface CalendarProps {
 
 export const MobileCalendar: React.FC<CalendarProps> = ({ calendar }) => {
   const { t } = useTranslation();
+
+  const StyledTableRow = styled(TableRow)(({ theme }) => ({
+    '&.current-event, &.next-event': {
+      backgroundColor: theme.palette.action.selected,
+      transition: 'background-color 200ms',
+    },
+  }));
 
   return (
     <TableContainer component={Paper} sx={{ padding: 2 }}>
@@ -22,14 +29,14 @@ export const MobileCalendar: React.FC<CalendarProps> = ({ calendar }) => {
         </TableHead>
         <TableBody>
           {calendar?.map((event) => (
-            <TableRow
+            <StyledTableRow
               key={JSON.stringify(event.date)}
               className={event.isCurrent ? 'current-event' : event.isNext ? 'next-event' : ''}
             >
               <TableCell>{formatToShort(event.date)}</TableCell>
               <TableCell>{event.leagueName ?? '-'}</TableCell>
               <TableCell>{t(`enums.event_types.${event.eventType}`, { day: event.round })}</TableCell>
-            </TableRow>
+            </StyledTableRow>
           ))}
         </TableBody>
       </Table>