+++ web-ui/src/components/dashboard-components/NewsContent.tsx
@@ -90,9 +90,9 @@ export const NewsContent = ({ newsItem }: NewsContentProps) => {
             <>
               <Typography variant="body1">{t('enums.news_type.NEW_SEASON.headline')}</Typography>
               <ul>
-                {(newsItem.retired as any[])?.map((b) => (
-                  <li key={`${b.id}-retired`}>
-                    <IconNameWithDetailsModal blob={b} blobId={b.id} name={b.name} color={b.color} />
+                {newsItem.retired?.map((b) => (
+                  <li key={`${b.name}-retired`}>
+                    <IconNameWithDetailsModal blob={b} name={b.name} color={b.color} />
                   </li>
                 ))}
               </ul>
@@ -107,9 +107,9 @@ export const NewsContent = ({ newsItem }: NewsContentProps) => {
                     {t('enums.news_type.NEW_SEASON.transfers', { leagueName: league.leagueName })}
                   </Typography>
                   <ul>
-                    {(league.blobs as any[]).map((b) => (
-                      <li key={`${b.id}-transfers`}>
-                        <IconNameWithDetailsModal blob={b} blobId={b.id} name={b.name} color={b.color} />
+                    {league.blobs.map((b) => (
+                      <li key={`${b.name}-transfers`}>
+                        <IconNameWithDetailsModal blob={b} name={b.name} color={b.color} />
                       </li>
                     ))}
                   </ul>
@@ -119,9 +119,9 @@ export const NewsContent = ({ newsItem }: NewsContentProps) => {
             <>
               <Typography variant="body1">{t('enums.news_type.NEW_SEASON.rookies')}</Typography>
               <ul>
-                {(newsItem.rookies as any[])?.map((b) => (
-                  <li key={`${b.id}-rookies`}>
-                    <IconNameWithDetailsModal blob={b} blobId={b.id} name={b.name} color={b.color} />
+                {newsItem.rookies?.map((b) => (
+                  <li key={`${b.name}-rookies`}>
+                    <IconNameWithDetailsModal blob={b} name={b.name} color={b.color} />
                   </li>
                 ))}
               </ul>