+++ web-ui/src/components/event-components/ResultsModal.tsx
@@ -1,14 +1,4 @@
-import {
-  Dialog,
-  DialogTitle,
-  DialogContent,
-  DialogActions,
-  Button,
-  CircularProgress,
-  Box,
-  Typography,
-  IconButton,
-} from '@mui/material';
+import { Dialog, DialogTitle, DialogContent, CircularProgress, Box, Typography, IconButton } from '@mui/material';
 import { useEffect, useState } from 'react';
 import { useMutation } from '@tanstack/react-query';
 import defaultConfig from '../../default-config';