+++ web-ui/src/components/icons/BlobPiktogram.tsx
@@ -1,19 +1,14 @@
-import { BlobIconProps } from "./BlobIcon";
+import { BlobIconProps } from './BlobIcon';
+import { useTheme } from '@mui/material/styles';
 
 export function BlobPiktogram({ size, color }: BlobIconProps) {
+  const theme = useTheme();
+  const eyeColor = theme.palette.mode === 'dark' ? '#000000' : '#ffffff';
   return (
-    <svg
-      width={size}
-      height={size}
-      viewBox="0 0 130 95"
-      version="1.1"
-      id="svg5">
-      <defs
-        id="defs2" />
-      <g
-        id="layer1">
-        <g
-          id="g12632">
+    <svg width={size} height={size} viewBox="0 0 130 95" version="1.1" id="svg5">
+      <defs id="defs2" />
+      <g id="layer1">
+        <g id="g12632">
           <path
             id="path56"
             fill={color}
@@ -21,11 +16,12 @@ export function BlobPiktogram({ size, color }: BlobIconProps) {
             fillRule="nonzero"
             stroke={color}
             strokeWidth="0.982221"
-            d="m 129.5089,59.689952 c 0,27.20034 -28.7716,34.818936 -64.263171,34.818937 -35.491565,0 -64.7546185,-7.618596 -64.7546185,-34.818937 C 0.49111025,32.48961 29.754163,0.4911092 65.245729,0.49110935 100.7373,0.49111061 129.5089,32.489611 129.5089,59.689952 Z" />
+            d="m 129.5089,59.689952 c 0,27.20034 -28.7716,34.818936 -64.263171,34.818937 -35.491565,0 -64.7546185,-7.618596 -64.7546185,-34.818937 C 0.49111025,32.48961 29.754163,0.4911092 65.245729,0.49110935 100.7373,0.49111061 129.5089,32.489611 129.5089,59.689952 Z"
+          />
           <ellipse
-            fill="#ffffff"
+            fill={eyeColor}
             fillOpacity="1"
-            stroke="#ffffff"
+            stroke={eyeColor}
             strokeWidth="5"
             strokeMiterlimit="4"
             strokeDasharray="none"
@@ -34,11 +30,12 @@ export function BlobPiktogram({ size, color }: BlobIconProps) {
             cx="45"
             cy="35"
             rx="4.4549451"
-            ry="9.4549446" />
+            ry="9.4549446"
+          />
           <ellipse
-            fill="#ffffff"
+            fill={eyeColor}
             fillOpacity="1"
-            stroke="#ffffff"
+            stroke={eyeColor}
             strokeWidth="5"
             strokeMiterlimit="4"
             strokeDasharray="none"
@@ -47,7 +44,8 @@ export function BlobPiktogram({ size, color }: BlobIconProps) {
             cx="85"
             cy="35"
             rx="4.4549451"
-            ry="9.4549446" />
+            ry="9.4549446"
+          />
         </g>
       </g>
     </svg>